# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SessionLayer Control Plane API
  version: 0.1.0
  summary: Contract-first REST surface for the SessionLayer Control Plane.
  description: |
    The **SessionLayer Control Plane API** is the contract-first source of truth
    for the platform's REST surface (Design §13, FR-API-1). The UI is a *client* of
    this contract; server interfaces and typed clients are code-generated from this
    document, and CI fails if generated code drifts from the spec. Postgres is the
    single source of truth; the UI and the API share one write path (D11).

    **Complete contract (Session 17).** This document is the complete, frozen
    resource surface: the config resources (`node-policies`, `rules`, `roles` +
    `role-bindings`, `cas`, `capability-defs`, `jit-policies`,
    `breakglass-policies`, `service-accounts`), the runtime resources (`nodes`,
    `sessions`, `locks`, `jit-requests`, `join-tokens`, `otp`, device-flow, `pins`,
    `breakglass`, `recordings`, `audit-events`), and the operational probes. The
    `recordings` replay/export and `audit-events` search operations are frozen here
    and return `501` until Session 18 implements them, so the Dashboard can be built
    in parallel against a finalized typed client.

    **Config vs runtime.** Config resources carry an `origin` provenance label
    (`api|ui|default`) and an optimistic-concurrency `version`; runtime resources
...
    API-only runtime resource (Design §13).

...
    compatibility policy.
...
servers:
...
security:
  - oidcBearer: []
  - clientCredentials: []
  - mtls: []
paths:
...
  /v1/locks:
    get:
      operationId: listLocks
      summary: List active locks (admin).
      description: |
        Lists the current unexpired locks. Platform-RBAC gated (`lock:read`) +
        audited (Design §8.3; FR-LOCK-1). A lock is the incident-response deny
        primitive; the deny reason is operator-only.
      tags:
        - locks
      responses:
        '200':
          description: The active locks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LockList'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createLock
      summary: Create a lock (admin).
      description: |
        Creates an incident-response lock that blocks new session issuance AND
        tears down existing matching sessions on every Gateway (Design §8.3,
        §8.4; FR-LOCK-1/2), pushed over the actively-pushed deny-list. Ingest
        validation rejects an empty/unrecognised target; a fleet-wide lock
        requires an explicit `all: true`. Platform-RBAC gated (`lock:write`) +
        audited.
      tags:
        - locks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLockRequest'
      responses:
        '201':
          description: The lock was created and pushed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LockResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/locks/{lockId}:
    delete:
      operationId: releaseLock
      summary: Release a lock (admin).
      description: |
        Releases (deletes) a lock; the removal is pushed to every Gateway.
        Releasing a lock only stops future denial — it never resurrects a
        torn-down session. Platform-RBAC gated (`lock:write`) + audited.
      tags:
        - locks
      parameters:
        - name: lockId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The lock was released (idempotent).
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC bearer access via an ID/JWT token validated by the Control Plane
        (Design §5.7, FR-AUTH-17). The ID token is the authentication proof.
    clientCredentials:
      type: oauth2
      description: |
        OAuth 2.0 client-credentials for machine consumers (Design §5.6,
        FR-AUTH-12). `private_key_jwt`/mTLS client authentication is preferred
        over static secrets.
      flows:
        clientCredentials:
          tokenUrl: https://controlplane.sessionlayer.example/oauth2/token
          scopes: {}
    mtls:
      type: mutualTLS
      description: |
        Mutual TLS client-certificate authentication (Design §5.7, FR-AUTH-17),
        also used by the internal CP<->Gateway plane.
...
  responses:
    Problem:
      description: An error occurred. The body is an RFC 9457 problem document.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
    ProblemDetails:
      type: object
      title: RFC 9457 Problem Details
      description: |
        Problem details for HTTP APIs as defined by RFC 9457 (obsoletes RFC 7807).
        Error responses use media type `application/problem+json`.
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          description: A URI reference identifying the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: A URI reference identifying the specific occurrence.
      additionalProperties: true
...
    LockTarget:
      type: object
      title: Lock Target
      description: |
        What the lock matches (S5 LockMatching facets; OR-matched). At least one
        facet must be non-empty, OR `all` must be true — a global lock is never
        implicit (ingest rejects an empty target).
      properties:
        identities:
          type: array
          items:
            type: string
          description: Resolved subject identities to deny.
        groups:
          type: array
          items:
            type: string
          description: SSO/OIDC groups to deny (membership match).
        nodeIds:
          type: array
          items:
            type: string
            format: uuid
          description: Target node ids to deny.
        principals:
          type: array
          items:
            type: string
          description: Linux logins (principals) to deny.
        nodeLabels:
          type: array
          items:
            type: string
          description: Node labels ("key=value") to deny.
        all:
          type: boolean
          description: An explicit fleet-wide lock (deny everything). Required for a target with no other facet.
      additionalProperties: false
    CreateLockRequest:
      type: object
      title: Create Lock Request
      required:
        - target
        - reason
      properties:
        target:
          $ref: '#/components/schemas/LockTarget'
        reason:
          type: string
          description: Operator/audit reason (mandatory). Never disclosed to the SSH user (§7.1).
        ttlSeconds:
          type: integer
          format: int64
          description: Optional lock lifetime in seconds; omitted or null = no expiry (indefinite until released).
      additionalProperties: false
    LockResource:
      type: object
      title: Lock
      required:
        - id
        - target
        - reason
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        target:
          $ref: '#/components/schemas/LockTarget'
        reason:
          type: string
        expiresAt:
          type: string
          format: date-time
          description: Absolute expiry; absent when the lock has no TTL.
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          description: The admin identity that created the lock.
      additionalProperties: false
    LockList:
      type: object
      title: Lock List
      required:
        - locks
      properties:
        locks:
          type: array
          items:
            $ref: '#/components/schemas/LockResource'
      additionalProperties: false
...